259: unified implementation of RelationsGraph and NetworkGraph
